+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 # Changelog
 
+## 0.15.0 (2023-07-22)
+
+- Add: Option to SO Test Page to override number of infotrons needed.
+- Fix: Produced MPX files now should work better in Megaplex. Dropped support of
+  incorrect MPX header (with spaces) when saving MPX file with 1 level.
+- Enh: Short options to SO Test Page
+
+### Desktop
+
+- Add: Logging errors to a log file for debug purpose to simplify feedback.
+- Add: Display app directories in "Info" dialog.
+
 ## 0.14.0 (2023-06-23)
 
 - Add: Demo Signature editing.